[feature #162170981] refactor signin form
