+++ src/component/ui/Modal/TriggerModal.scss
@@ -1,3 +1,7 @@
 .modal-header {
   border-bottom: none;
 }
+
+.modal-link {
+ background-color: antiquewhite
+}
