+++ src/component/ui/errors/Signuperror.spec.js
@@ -2,11 +2,11 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import SignInForm from './SignInForm';
+import SignupError from './signupError';
 
 describe('The Signup Component Test Suite', () => {
   it('should match the snapshot', () => {
-    const tree = shallow(<SignInForm />);
+    const tree = shallow(<SignupError />);
     expect(toJson(tree)).toMatchSnapshot();
   });
 });
